requirements.txt
# XenusAI Dependencies
# Install: pip install -r requirements.txt

# ── Embeddings & LLM ──
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.35.0

# ── Vector Store ──
chromadb>=0.4.0

# ── Text Processing ──
beautifulsoup4>=4.12.0
trafilatura>=1.6.0
tiktoken>=0.5.0
lxml>=4.9.0

# ── Web Requests ──
requests>=2.31.0

# ── GUI ──
eel>=0.16.0

# ── CLI & Display ──
rich>=13.0.0

# ── File Handling ──
chardet>=5.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11

# ── Voice (STT & TTS) ──
sounddevice>=0.5.0
soundfile>=0.12.0
SpeechRecognition>=3.10.0
numpy>=1.24.0
scipy>=1.10.0

# ── Auto-Learning ──
wikipedia>=1.4.0
